+++ src/openscm_runner/adapters/magicc7/_parallel_process.py
@@ -5,13 +5,9 @@ import logging
 import time
 from concurrent.futures import as_completed
 
-from tqdm.autonotebook import tqdm
+from ...progress import progress
 
 LOGGER = logging.getLogger(__name__)
-
-
-# TODO: handle configuration in a consistent manner
-_default_tqdm_params = {"mininterval": 5}
 
 
 def _run_serial(func, configs, config_are_kwargs, desc):
@@ -19,10 +15,10 @@ def _run_serial(func, configs, config_are_kwargs, desc):
 
     if config_are_kwargs:
         LOGGER.debug("Treating config as kwargs")
-        res = [func(**a) for a in tqdm(configs, desc=desc, **_default_tqdm_params)]
+        res = [func(**a) for a in progress(configs, desc=desc)]
     else:
         LOGGER.debug("Treating config as args")
-        res = [func(a) for a in tqdm(configs, desc=desc, **_default_tqdm_params)]
+        res = [func(a) for a in progress(configs, desc=desc)]
 
     LOGGER.debug("Exiting _run_serial")
     return res
@@ -40,17 +36,9 @@ def _run_parallel(  # pylint:disable=too-many-arguments
         LOGGER.debug("Treating config as args")
         futures = [pool.submit(func, a) for a in configs]
 
-    kwargs_tqdm = {
-        **_default_tqdm_params,
-        "total": len(futures),
-        "unit": "it",
-        "unit_scale": True,
-        "desc": desc,
-    }
-
     LOGGER.debug("Waiting for jobs to complete")
-    for i, future in tqdm(
-        enumerate(as_completed(futures, timeout=timeout)), **kwargs_tqdm
+    for i, future in progress(
+        enumerate(as_completed(futures, timeout=timeout)), total=len(futures), desc=desc
     ):
         if future.exception() is not None:
             time.sleep(2)  # let buffer flush out